router: fix duplicate market bug.
Fix a bug where if a market is registered with a name, and the market is no longer connected, we don't allow markets to use that name in the future. now markets will be able to connect as long as the name is never used before or the market(s) that used it are no longer connected.